Finished basic AJAX requests, started excel exports
application.php now sends proper AJAX requests to application_form.php with no CORS issues- this was accomplished by using a constant session after the user signs in the first time through CAS in order to check their account's details. application.php receives an array of errors for anything that wasn't validated, but it still needs to render this list to the view.

Excel exports were experimented with; the goal is to use the FileSaver.js library to support javascript-based file downloads, so that custom application lists can be generated at will. Currently there are problems getting the FileSaver modules imported correctly to be used.