Added global site warning feature, improved database import files
Added a feature to create a site warning, which is viewable to all users while visible. This is saved as a 'SiteWarning' variable in the database under the 'variables' table.

Updated the database shema import file, and moved it to the /sql directory. Also created a data.sql file for importing static data (including the variable mentioned above)