Added page footer, fixed some small bugs
All pages now include a footer, which holds contacts information, address information, and a copyright notice.

Fixed error checks in API to check the value of retval["success"] rather than the existance or retval["error"], since its possible for error messages not to appear. Also fixed EmailHelper to initialize sendSuccess to false as a failsafe.